+++ packages/mobile/src/screens/Channel/Channel.screen.tsx
@@ -51,9 +51,12 @@ export const ChannelScreen: FC = () => {
   const isWebsocketConnected = useSelector(initSelectors.isWebsocketConnected)
 
   let contextMenu: UseContextMenuType<Record<string, unknown>> | null = useContextMenu(MenuName.Channel)
+
   if (!community?.CA || !isWebsocketConnected) {
     contextMenu = null
   }
+
+  const unregisteredUsernameContextMenu = useContextMenu(MenuName.UnregisteredUsername)
 
   const [uploadingFiles, setUploadingFiles] = React.useState<FilePreviewData>({})
   const filesRef = React.useRef<FilePreviewData>({})
@@ -119,16 +122,9 @@ export const ChannelScreen: FC = () => {
 
   const unregisteredUsernameHandleBack = useCallback(
     (username: string) => {
-      dispatch(
-        navigationActions.navigation({
-          screen: ScreenNames.UnregisteredUsernameScreen,
-          params: {
-            username,
-          },
-        })
-      )
+      unregisteredUsernameContextMenu.handleOpen({ username })
     },
-    [dispatch]
+    [unregisteredUsernameContextMenu]
   )
 
   const sendMessageAction = React.useCallback(